requirements.txt
aiocontextvars==0.2.2
aiohttp==2.3.10
async-timeout==2.0.1
burnysc2==5.0.0
chardet==3.0.4
colorama==0.4.3
contextvars==2.4
idna==2.10
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==1.7.0
loguru==0.5.1
mpyq==0.2.5
multidict==4.7.6
numpy==1.18.5
portpicker==1.3.1
protobuf==3.12.4
pyglet==1.5.7
s2clientprotocol==5.0.2.81102.0
scipy==1.4.1
six==1.15.0
typing-extensions==3.7.4.2
win32-setctime==1.0.1
yarl==1.5.1
zipp==1.2.0
